Improved test bench
This version is based off of ood-images-full. It includes an Ondemand
installation as well as a Slurm installation. Future work should look
into removing the expect scripts and switching to gridengine's conf
scripts. Likewise the current installation settings do not appear to use
the head node as a batch node.